Rearrange integration tests
